Enable 64-bit kernel boot

+++ src/kernel/mm/mm.c
@@ -5,14 +5,14 @@
 
 extern uint8_t end; /* Symbol defined by linker */
 
-static uint32_t next_free_page;
+static uint64_t next_free_page;
 
 /*
  * Initialize a very small physical page allocator. The allocator simply
  * hands out pages sequentially starting after the kernel image.
  */
 void mm_init(void) {
-    next_free_page = ((uint32_t)&end + PAGE_SIZE - 1) & ~(PAGE_SIZE - 1);
+    next_free_page = ((uint64_t)&end + PAGE_SIZE - 1) & ~(PAGE_SIZE - 1);
 }
 
 void *page_alloc(void) {
